Added logback logging and replaced System.{out,err} with logging statements. Removed JSON sample classes that are unrelated to the Fractals project. Moved JsonReaderWriter class to com.sarahhand.json.
