--- src/main/java/com/sarahhand/apps/App.java
@@ -1,14 +1,0 @@
-package com.sarahhand.apps;
-
-public class App{
-	
-	private String name;
-	
-    public App(String name){
-    	this.name = name;
-    }
-    
-	public String getName(){
-		return name;
-	}
-}
