--- src/main/java/com/sarahhand/apps/model/Address.java
@@ -1,46 +1,0 @@
-package com.sarahhand.apps.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-
-//@JsonIgnoreProperties(ignoreUnknown=true)
-@JsonInclude(Include.NON_NULL)
-public class Address {
-	
-	private String type;
-	private String street1;
-	private String city;
-	private String state;
-	private String zipCode;
-	
-	public String getType() {
-		return type;
-	}
-	public void setType(String type) {
-		this.type = type;
-	}
-	public String getStreet1() {
-		return street1;
-	}
-	public void setStreet1(String street1) {
-		this.street1 = street1;
-	}
-	public String getCity() {
-		return city;
-	}
-	public void setCity(String city) {
-		this.city = city;
-	}
-	public String getState() {
-		return state;
-	}
-	public void setState(String state) {
-		this.state = state;
-	}
-	public String getZipCode() {
-		return zipCode;
-	}
-	public void setZipCode(String zipcode) {
-		this.zipCode = zipcode;
-	}
-}
