--- src/main/java/com/sarahhand/apps/model/Contact.java
@@ -1,48 +1,0 @@
-package com.sarahhand.apps.model;
-
-import java.util.Set;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-
-//@JsonIgnoreProperties(ignoreUnknown=true)
-@JsonInclude(Include.NON_NULL)
-public class Contact {
-	
-	private String firstName;
-	private String lastName;
-	private String middleName;
-	private Set<PhoneNumber> phoneNumbers;
-	private Set<Address> addresses;
-	
-	public String getFirstName() {
-		return firstName;
-	}
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-	public String getLastName() {
-		return lastName;
-	}
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-	public Set<PhoneNumber> getPhoneNumbers() {
-		return phoneNumbers;
-	}
-	public void setPhoneNumbers(Set<PhoneNumber> phoneNumbers) {
-		this.phoneNumbers = phoneNumbers;
-	}
-	public Set<Address> getAddresses() {
-		return addresses;
-	}
-	public void setAddresses(Set<Address> addresses) {
-		this.addresses = addresses;
-	}
-	public String getMiddleName(){
-		return middleName;
-	}
-	public void setMiddleName(String middleName){
-		this.middleName = middleName;
-	}
-}
