--- src/main/java/com/sarahhand/apps/model/PhoneNumber.java
@@ -1,25 +1,0 @@
-package com.sarahhand.apps.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-
-//@JsonIgnoreProperties(ignoreUnknown=true)
-@JsonInclude(Include.NON_NULL)
-public class PhoneNumber {
-	
-	private String type;
-	private String number;
-	
-	public String getType() {
-		return type;
-	}
-	public void setType(String type) {
-		this.type = type;
-	}
-	public String getNumber() {
-		return number;
-	}
-	public void setNumber(String number) {
-		this.number = number;
-	}
-}
