+++ src/main/java/com/sarahhand/fractals/model/MandelbrotViewerImpl.java
@@ -9,6 +9,9 @@ import java.awt.Rectangle;
 import java.awt.geom.Point2D.Double;
 import java.awt.image.BufferedImage;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 /**
  * Implements <code>MandelbrotViewer</code>.
  * 
@@ -17,6 +20,7 @@ import java.awt.image.BufferedImage;
  */
 
 class MandelbrotViewerImpl implements MandelbrotViewer{
+	private final Logger log = LoggerFactory.getLogger(this.getClass());
 	
 	private static final double LOG2 = Math.log(2);
 	private static final int MIN_SIZE = 10;
@@ -44,7 +48,7 @@ class MandelbrotViewerImpl implements MandelbrotViewer{
 		
 		renderRect(g, new Rectangle(dimensions), dimensions, MIN_SIZE);
 		
-		System.out.println("Time: " + ((System.currentTimeMillis()-time)/1000) + "(secs)");
+		log.debug("Time to generate Mandelbrot image: {}ms.", System.currentTimeMillis()-time);
 		return image;
 	}
 	
@@ -180,7 +184,7 @@ class MandelbrotViewerImpl implements MandelbrotViewer{
 		return new Color((int)r, (int)g, (int)b);
 	}
 	
-	private class ComplexNumber{
+	private static class ComplexNumber{
 		
 		double x;
 		double y;
