+++ src/main/java/com/sarahhand/json/JsonReaderWriter.java
@@ -1,8 +1,11 @@
-package com.sarahhand.apps;
+package com.sarahhand.json;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -10,6 +13,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.SerializationFeature;
 
 public class JsonReaderWriter{
+	private final Logger log = LoggerFactory.getLogger(this.getClass());
 	
 	private ObjectMapper mapper = new ObjectMapper();
 	
@@ -20,13 +24,12 @@ public class JsonReaderWriter{
 		mapper.setPropertyNamingStrategy(pns);
 	}
 	
-	public <T> T read(String fileName, Class<? extends T> target) {
-		try (InputStream stream = this.getClass().getResourceAsStream(fileName)){
+	public <T> T read(InputStream stream, Class<? extends T> target) {
+		try {
 			T object = mapper.readValue(stream, target);
 			return object;
 		} catch (IOException e) {
-			System.err.println("Error reading contact from file=" + fileName);
-			e.printStackTrace();
+			log.error("Error reading JSON from stream.", e);
 		}
 		return null;
 	}
@@ -35,8 +38,7 @@ public class JsonReaderWriter{
 		try{
 			mapper.writeValue(output, object);
 		} catch (IOException e){
-			System.err.println("Error writing contact to file");
-			e.printStackTrace();
+			log.error("Error writing JSON to stream.", e);
 		}
 	}
 }
