--- src/test/resources/com/sarahhand/apps/Contact.json
@@ -1,24 +1,0 @@
-{
-	"first_name" : "Roger",
-	"last_name" : "Rabbit",
-	"phone_numbers" : [
-		{
-			"type" : "home",
-		 	"number" : "212-555-1212"
-		},
-		{
-			"type" : "mobile",
-		 	"number" : "212-555-2323"
-		}
-	],
-	"addresses" : [
-		{
-			"type" : "home",
-			"street1" : "100 Main St.",
-			"street2" : "Apt. 2",
-			"city" : "Bellevue",
-			"state" : "WA",
-			"zip_code" : "98004"
-		}
-	]
-}
